Moved NotFoundHandler to ErrorHandler namespace
